Restructure:Create new sub dirs for components... - components/my-classes/ - components/new-class
