+++ components/new-class/AddNewStudent.tsx
